requirements.txt
# Core dependencies
python-dotenv==1.0.0
pinecone>=6.0.0
streamlit==1.32.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
tqdm==4.66.2
loguru==0.7.2
pycryptodome==3.22.0
PyMuPDF==1.23.3
tiktoken==0.9.0

# Web API
fastapi==0.109.2
uvicorn==0.27.1
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
sentence-transformers>=2.2.2
numpy>=1.24.0
loguru>=0.7.0
tabulate==0.9.0 